we need to call `interpret` at comptime.
1. we're able to run hardcoded brainfuck programs with hardcoded input
   at comptime (if the comptime evaluation doesn't take too many steps to
   make compiler emit an error)
2. hardcoded program is fine, but we don't want to hardcode the input.

in all, why this current approach doesn't work for custom user input?

because ',' requires `stdin.readByte()` which is an extern call and thus
doesn't work at comptime. similarly for '.'. but we don't want to input
and output stuff at comptime, so we need a new approach anyways.

as a consequence, we can't just throw a large `comptime` block to
contain `interpret`, but like you can if it doesn't involve inputting and
outputting, see (sidenote).

the obivous step is to `inline while` the interpreter loop. but it
doesn't work since in the condition `ip != prog.len`, `ip` is not known
at comptime.

sidenote: normally you can't call extern functions at comptime, but if
you are in a `comptime` block but the extern call is not actually in the
control path, in which case zig will not complain, you're fine.